Adjust messages.sh to not produce errors when no .rc or .ui files are found.

+++ plugins/Messages.sh
@@ -1,8 +1,25 @@
 #! /usr/bin/env bash
+rcfiles="`find $subdirs -name \*.rc`"
+uifiles="`find $subdirs -name \*.ui`"
+if [[ "$rcfiles" != "" ]] ; then
+    $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+fi
+if [[ "$uifiles" != "" ]] ; then
+    $EXTRACTRC  >> rc.cpp || exit 12
+fi
+$XGETTEXT -kaliasLocal `find $subdirs -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdevelop.pot
+rm -f rc.cpp
+
 plugins="appwizard coverage cppdebugger documentview formatters grepview qtdesigner valgrind xtest"
 for plugin in $plugins ; do
-    $EXTRACTRC `find $plugin -name \*.rc` >> rc.cpp || exit 11
-    $EXTRACTRC `find $plugin -name \*.ui` >> rc.cpp || exit 12
+    rcfiles="`find $plugin -name \*.rc`"
+    uifiles="`find $plugin -name \*.ui`"
+    if [[ "$rcfiles" != "" ]] ; then
+        $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+    fi
+    if [[ "$uifiles" != "" ]] ; then
+        $EXTRACTRC  >> rc.cpp || exit 12
+    fi
     $XGETTEXT -kaliasLocal `find $plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
     rm -f rc.cpp
 done
@@ -11,24 +28,42 @@ done
 
 builders="qmakebuilder cmakebuilder makebuilder"
 for plugin in $builders ; do
-    $EXTRACTRC `find builders/$plugin -name \*.rc` >> rc.cpp || exit 11
-    $EXTRACTRC `find builders/$plugin -name \*.ui` >> rc.cpp || exit 12
+    rcfiles="`find $plugin -name \*.rc`"
+    uifiles="`find $plugin -name \*.ui`"
+    if [[ "$rcfiles" != "" ]] ; then
+        $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+    fi
+    if [[ "$uifiles" != "" ]] ; then
+        $EXTRACTRC  >> rc.cpp || exit 12
+    fi
     $XGETTEXT -kaliasLocal `find builders/$plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
     rm -f rc.cpp
 done
 
 managers="qmake cmake custommake"
 for plugin in $managers ; do
-    $EXTRACTRC `find managers/$plugin -name \*.rc` >> rc.cpp || exit 11
-    $EXTRACTRC `find managers/$plugin -name \*.ui` >> rc.cpp || exit 12
+    rcfiles="`find $plugin -name \*.rc`"
+    uifiles="`find $plugin -name \*.ui`"
+    if [[ "$rcfiles" != "" ]] ; then
+        $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+    fi
+    if [[ "$uifiles" != "" ]] ; then
+        $EXTRACTRC  >> rc.cpp || exit 12
+    fi
     $XGETTEXT -kaliasLocal `find managers/$plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
     rm -f rc.cpp
 done
 
 languages="cpp"
 for plugin in $languages ; do
-    $EXTRACTRC `find languages/$plugin -name \*.rc` >> rc.cpp || exit 11
-    $EXTRACTRC `find languages/$plugin -name \*.ui` >> rc.cpp || exit 12
+    rcfiles="`find $plugin -name \*.rc`"
+    uifiles="`find $plugin -name \*.ui`"
+    if [[ "$rcfiles" != "" ]] ; then
+        $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+    fi
+    if [[ "$uifiles" != "" ]] ; then
+        $EXTRACTRC  >> rc.cpp || exit 12
+    fi
     $XGETTEXT -kaliasLocal `find languages/$plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
     rm -f rc.cpp
 done
